requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# LLM and API dependencies
openai>=1.0.0
zhipuai>=2.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0

# Async and HTTP
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Environment and configuration
python-dotenv>=0.19.0
pydantic>=1.8.0

# Utilities
tqdm>=4.62.0
colorama>=0.4.4

# Development dependencies (optional)
pytest>=6.2.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910